+++ public_test.go
@@ -10,11 +10,7 @@ import (
 
 func TestNew(t *testing.T) {
 	// Create a real folder
-	ok_dir, err := ioutil.TempDir("/tmp", "smugmug-backup")
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer os.RemoveAll(ok_dir)
+	ok_dir := t.TempDir()
 
 	// Create and immediately delete a folder, so we are sure it doesn't exist
 	unexisting_dir, err := ioutil.TempDir("/tmp", "smugmug-backup")
@@ -25,7 +21,6 @@ func TestNew(t *testing.T) {
 
 	tests := []struct {
 		name        string
-		username    string
 		destination string
 		apiKey      string
 		apiSecret   string
@@ -35,7 +30,6 @@ func TestNew(t *testing.T) {
 	}{
 		{
 			name:        "empty name",
-			username:    "",
 			destination: "/path/to/dest/",
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -45,7 +39,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "empty destination",
-			username:    "user_name",
 			destination: "",
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -55,7 +48,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "unexisting dest dir",
-			username:    "user_name",
 			destination: unexisting_dir,
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -65,7 +57,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "missing apiKey",
-			username:    "user_name",
 			destination: ok_dir,
 			apiKey:      "",
 			apiSecret:   "value",
@@ -75,7 +66,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "missing apiSecret",
-			username:    "user_name",
 			destination: ok_dir,
 			apiKey:      "value",
 			apiSecret:   "",
@@ -85,7 +75,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "missing userToken",
-			username:    "user_name",
 			destination: ok_dir,
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -95,7 +84,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "missing userSecret",
-			username:    "user_name",
 			destination: ok_dir,
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -105,7 +93,6 @@ func TestNew(t *testing.T) {
 		},
 		{
 			name:        "correct",
-			username:    "user_name",
 			destination: ok_dir,
 			apiKey:      "value",
 			apiSecret:   "value",
@@ -118,7 +105,6 @@ func TestNew(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			_, err := smugmug.New(&smugmug.Conf{
-				Username:    tt.username,
 				Destination: tt.destination,
 				ApiKey:      tt.apiKey,
 				ApiSecret:   tt.apiSecret,